% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probability.R
\name{P}
\alias{P}
\title{probability of an outcome/outcomes.}
\usage{
P(x)
}
\arguments{
\item{x}{a random variable}
}
\value{
a probability, a number between 0 and 1.

a numeric vector of probabilities.
}
\description{
probability of an outcome/outcomes.
}
\examples{
dice <- rv(1:6)
P(dice == 2)
}

